Extract shared integration test helpers
